+++ lib/components/ChatView/index.d.ts
@@ -4,8 +4,9 @@ import React from 'react';
 declare interface ChatViewProps {
     renderRow: (item: any, index: number, data: Array) => React.ReactElement<any>;
     onFetch?: (page: number, fill: (data: any, allLoaded: boolean) => void) => void;
-    style?: object,
-    allLoadedText: string
+    style?: object;
+    allLoadedText?: string;
+    stayPosition?: boolean;
 }
 
 declare class ChatView extends React.Component<ChatViewProps, any> {